refactor(breaking): rename breaking module + dual-vocabulary change-category aliasing
M4 (OSS half) of the change-category vocabulary migration. Opens the
dual-vocabulary aliasing window: wire enums stay legacy
(breaking/non_breaking/partial_breaking/unknown); the v2 vocabulary
(model_wide/column/additive) is accepted on input and emitted opt-in via the
`Accept-Vocabulary: v2` response header.

- Rename recce/util/breaking.py -> recce/util/change_classifier.py;
  breaking.py kept as a deprecated re-export shim (removed next release).
- Add CHANGE_CATEGORY_ALIASES + normalize_change_category() +
  to_v2_change_category() (in models/types.py to avoid a circular import,
  re-exported from change_classifier).
- Defensive Pydantic field validators normalize v2 input to legacy.
- Opt-in Accept-Vocabulary: v2 output on /api/info and /api/cll.
- CI lint (scripts/check_wire_enum_literals.sh + workflow) blocks NEW legacy
  wire-enum literals under js/packages/ui/src/; allowlist marker // wire-enum-ok.

Resolves DRC-3553 (OSS portion). Cloud-side is a separate repo/PR.
